fix: avoid bitshifting uint8

+++ Coder Desktop/VPNLib/Receiver.swift
@@ -90,5 +90,5 @@ func deserializeLen(_ data: Data) throws -> UInt32 {
     if data.count != 4 {
         throw ReceiveError.invalidLength
     }
-    return UInt32(data[0] << 24 | data[1] << 16 | data[2] << 8 | data[3])
+    return UInt32(data[0]) << 24 | UInt32(data[1]) << 16 | UInt32(data[2]) << 8 | UInt32(data[3])
 }
